+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Community maps: Added "Requires 2 accounts" tag (with that tag you can make fair use of 4 mass destructions)
 * Community maps: Added ability to filter by author
 * Community maps: Fixed an issue where filters wouldn't work in "my maps"
+* Community maps: Added the ability to add custom icons to the map
 
 [Full Changelog](https://github.com/kabamgamer/dd2-strategy-hub/compare/v1.7.1...main)